fix(viewer/frontend): fixes aspect ratio refresh and camera NaN position on ze
this was a really fustrating debug session: because we're setting `display:none` on the renderer dom
element if the vue renderer comp is destroyed (to keep one instance alive always), this would
trigger the on window resize function. In turn, this would set the camera aspect ratio from a div
thats 0x0px. Needless to say, things would refuse to work thereafterwards and a full page refresh
would be needed to reinstantiate the renderer.